- Merging Kits

+++ tapuilibrary/src/main/java/company/tap/tapuilibrary/themekit/theme/BaseTextTheme.kt
@@ -1,11 +1,8 @@
 package company.tap.tapuilibrary.themekit.theme
 
-/**
- * Created by Mario Gamal on 4/20/20
- * Copyright © 2020 Tap Payments. All rights reserved.
- */
+
 open class BaseTextTheme {
     var textColor: Int? = null
-    var textSize: Float? = null
-    var letterSpacing: Float? = null
+    var textSize: Double? = null
+    var letterSpacing: Double? = null
 }
